Fix bug with network inteface name clashes
Before, the veth pair was created in the default network namespace, and
the Pod-end of the pair is always called eth0. This could cause name
clashes when the plugin was invoked multiple times in parallel, and a
new veth pair is created before the Pod-end of the previous one is moved
to the Pod network namespace.

Now, the veth pair is created in the Pod network namespace and the
host-end then moved to the default network namespace.