Fix one-time setup
1. Critical section with mutex

The entire one-time setup is placed in a critical section that is
protected with a mutex (flock). This ensures that at most one process
can do the one-time setup (this is necessary because the kubelet invokes
the CNI plugin in parallel if multiple Pods are created at the same
time).

At the same time, the mutex provides synchronisation. No process can
reach the Pod-specific setup before the one-time setup has been
completed at least once (if, for example, a process would just skip the
one-time setup if another processes signalises that it is currently
doing it).

2. Whether setup has been already done is checked for each command

Instead of a single flag that signalises whether the entire one-time
setup has been already done, this check is now done for each command. In
other words, the one-time setup is executed on every invocation of the
plugin, but it checks for each setting whether it has to be done or not.